# Reduce latency of conditional breakpoints/stop-resume cycles {WIP]