--- a/pkg/proc/core/core.go
+++ b/pkg/proc/core/core.go
@@ -155,10 +155,9 @@ type Process struct {
 
 	entryPoint uint64
 
-	bi                *proc.BinaryInfo
-	breakpoints       proc.BreakpointMap
-	currentThread     *Thread
-	selectedGoroutine *proc.G
+	bi            *proc.BinaryInfo
+	breakpoints   proc.BreakpointMap
+	currentThread *Thread
 }
 
 // Thread represents a thread in the core file being debugged.
@@ -212,29 +211,12 @@ func OpenCore(corePath, exePath string, debugInfoDirs []string) (*proc.Target, e
 		return nil, err
 	}
 
-	if err := p.initialize(exePath, debugInfoDirs); err != nil {
-		return nil, err
-	}
-
-	return proc.NewTarget(p), nil
-}
-
-// initialize for core files doesn't do much
-// aside from call the post initialization setup.
-func (p *Process) initialize(path string, debugInfoDirs []string) error {
-	return proc.PostInitializationSetup(p, path, debugInfoDirs, p.writeBreakpoint)
+	return proc.NewTarget(p, exePath, debugInfoDirs, p.writeBreakpoint)
 }
 
 // BinInfo will return the binary info.
 func (p *Process) BinInfo() *proc.BinaryInfo {
 	return p.bi
-}
-
-// SetSelectedGoroutine will set internally the goroutine that should be
-// the default for any command executed, the goroutine being actively
-// followed.
-func (p *Process) SetSelectedGoroutine(g *proc.G) {
-	p.selectedGoroutine = g
 }
 
 // EntryPoint will return the entry point address for this core file.
@@ -443,38 +425,9 @@ func (p *Process) Pid() int {
 func (p *Process) ResumeNotify(chan<- struct{}) {
 }
 
-// SelectedGoroutine returns the current active and selected
-// goroutine.
-func (p *Process) SelectedGoroutine() *proc.G {
-	return p.selectedGoroutine
-}
-
 // SetBreakpoint will always return an error for core files as you cannot write memory or control execution.
 func (p *Process) SetBreakpoint(addr uint64, kind proc.BreakpointKind, cond ast.Expr) (*proc.Breakpoint, error) {
 	return nil, ErrWriteCore
-}
-
-// SwitchGoroutine will change the selected and active goroutine.
-func (p *Process) SwitchGoroutine(g *proc.G) error {
-	if g == nil {
-		// user specified -1 and selectedGoroutine is nil
-		return nil
-	}
-	if g.Thread != nil {
-		return p.SwitchThread(g.Thread.ThreadID())
-	}
-	p.selectedGoroutine = g
-	return nil
-}
-
-// SwitchThread will change the selected and active thread.
-func (p *Process) SwitchThread(tid int) error {
-	if th, ok := p.Threads[tid]; ok {
-		p.currentThread = th
-		p.selectedGoroutine, _ = proc.GetG(p.CurrentThread())
-		return nil
-	}
-	return fmt.Errorf("thread %d does not exist", tid)
 }
 
 // ThreadList will return a list of all threads currently in the process.
@@ -490,4 +443,9 @@ func (p *Process) ThreadList() []proc.Thread {
 func (p *Process) FindThread(threadID int) (proc.Thread, bool) {
 	t, ok := p.Threads[threadID]
 	return t, ok
+}
+
+// InternalSetCurrentThread is used internally by proc.Target to change the current thread.
+func (p *Process) InternalSetCurrentThread(th proc.Thread) {
+	p.currentThread = th.(*Thread)
 }