--- a/pkg/proc/fbsdutil/regs.go
+++ b/pkg/proc/fbsdutil/regs.go
@@ -132,11 +132,6 @@ func (r *AMD64Registers) BP() uint64 {
 	return uint64(r.Regs.Rbp)
 }
 
-// CX returns the value of RCX register.
-func (r *AMD64Registers) CX() uint64 {
-	return uint64(r.Regs.Rcx)
-}
-
 // TLS returns the address of the thread local storage memory segment.
 func (r *AMD64Registers) TLS() uint64 {
 	return r.Fsbase