… pkg/proc/core/core_test.go
… pkg/proc/native/proc.go
… pkg/proc/native/proc_darwin.go
… pkg/proc/native/proc_freebsd.go
… pkg/proc/native/proc_linux.go
… pkg/proc/native/proc_windows.go
… pkg/proc/proc.go
… pkg/proc/proc_test.go
… pkg/proc/registers.go
… pkg/proc/stack.go
… pkg/proc/target.go
… pkg/proc/threads.go
… pkg/proc/variables.go
… pkg/proc/winutil/regs.go
… service/api/conversions.go
… service/debugger/debugger.go